chore(utxo-lib): move coinselect and compose related types
