+++ packages/utxo-lib/src/coinselect/index.ts
@@ -1,61 +1,8 @@
 import { accumulative } from './inputs/accumulative';
 import { bnb } from './inputs/bnb';
-import { sortByScore, anyOf, TxType } from './utils';
+import { sortByScore, anyOf } from './utils';
 import { tryConfirmed } from './tryconfirmed';
-
-export type CoinSelectOptions = {
-    txType: TxType;
-    dustThreshold: number;
-    own?: number;
-    other?: number;
-    coinbase?: number;
-    baseFee?: number;
-    floorBaseFee?: boolean;
-    dustOutputFee?: number;
-    skipPermutation?: boolean;
-};
-
-export type CoinSelectInput = {
-    type: TxType;
-    i: number;
-    script: { length: number };
-    value: string;
-    confirmations: number;
-    coinbase?: boolean;
-    required?: boolean;
-    own?: boolean;
-    weight?: number;
-};
-
-export type CoinSelectOutput = {
-    script: { length: number };
-    value?: string;
-    weight?: number;
-};
-
-export type CoinSelectOutputFinal = {
-    script: { length: number };
-    value: string;
-};
-
-export type CoinSelectResult =
-    | {
-          fee: number;
-          inputs?: typeof undefined;
-          outputs?: typeof undefined;
-      }
-    | {
-          fee: number;
-          inputs: CoinSelectInput[];
-          outputs: CoinSelectOutputFinal[];
-      };
-
-export type CoinSelectAlgorithm = (
-    inputs: CoinSelectInput[],
-    outputs: CoinSelectOutput[],
-    feeRate: number,
-    options: CoinSelectOptions,
-) => CoinSelectResult;
+import { CoinSelectInput, CoinSelectOutput, CoinSelectOptions } from '../types';
 
 export function coinselect(
     inputs: CoinSelectInput[],
