+++ packages/utxo-lib/src/coinselect/outputs/split.ts
@@ -5,7 +5,7 @@ import type {
     CoinSelectOutput,
     CoinSelectOptions,
     CoinSelectResult,
-} from '../index';
+} from '../../types';
 
 // split utxos between each output, ignores outputs with .value defined
 export function split(
