+++ packages/utxo-lib/src/compose/result.ts
@@ -1,40 +1,12 @@
-import type { ComposedTransaction } from './transaction';
-import type { CompleteResult } from './coinselect';
-
-// Output from coinselect algorithm
-// 'nonfinal' - contains info about the outputs, but not Trezor tx
-// 'final' - contains info about outputs + Trezor tx
-// 'error' - some error, so far only NOT-ENOUGH-FUNDS and EMPTY strings
-export type ComposeResult =
-    | {
-          type: 'error';
-          error: string;
-      }
-    | {
-          type: 'nonfinal';
-          max?: string;
-          totalSpent: string; // all the outputs, no fee, no change
-          fee: string;
-          feePerByte: string;
-          bytes: number;
-      }
-    | {
-          type: 'final';
-          max?: string;
-          totalSpent: string; // all the outputs, no fee, no change
-          fee: string;
-          feePerByte: string;
-          bytes: number;
-          transaction: ComposedTransaction;
-      };
+import type { CoinSelectSuccess, ComposedTransaction, ComposeResult } from '../types';
 
 export const empty: ComposeResult = {
     type: 'error',
     error: 'EMPTY',
 };
 
-export function getNonfinalResult(result: CompleteResult): ComposeResult {
-    const { max, fee, feePerByte, bytes, totalSpent } = result.result;
+export function getNonfinalResult(result: CoinSelectSuccess): ComposeResult {
+    const { max, fee, feePerByte, bytes, totalSpent } = result.payload;
 
     return {
         type: 'nonfinal',
@@ -47,10 +19,10 @@ export function getNonfinalResult(result: CompleteResult): ComposeResult {
 }
 
 export function getFinalResult(
-    result: CompleteResult,
+    result: CoinSelectSuccess,
     transaction: ComposedTransaction,
 ): ComposeResult {
-    const { max, fee, feePerByte, bytes, totalSpent } = result.result;
+    const { max, fee, feePerByte, bytes, totalSpent } = result.payload;
 
     return {
         type: 'final',
