+++ packages/utxo-lib/src/types/index.ts
@@ -1,2 +1,4 @@
+export * from './coinselect';
+export * from './compose';
 export * from './payments';
 export * from './typeforce';
