+++ packages/utxo-lib/tests/compose.utils.ts
@@ -7,9 +7,8 @@
 // https://github.com/trezor/trezor-firmware/blob/1fceca73da523c5bf2bb0f398c91e00c728bdbe0/core/tests/test_apps.bitcoin.txweight.py
 
 import * as baddress from '../src/address';
-import { OUTPUT_SCRIPT_LENGTH, TxType } from '../src/coinselect/utils';
-import type { Network } from '../src';
-import type { ComposeResult } from '../src/compose/result';
+import { OUTPUT_SCRIPT_LENGTH } from '../src/coinselect/utils';
+import type { Network, ComposeResult, CoinSelectPaymentType } from '../src';
 
 // transaction header size: 4 byte version
 const _TXSIZE_HEADER = 4;
@@ -188,7 +187,7 @@ type FinalResult = Exclude<ComposeResult, { type: 'nonfinal' } | { type: 'error'
 
 export function verifyTxBytes(
     tx: FinalResult,
-    txType: Exclude<TxType, 'p2wsh'> = 'p2pkh',
+    txType: Exclude<CoinSelectPaymentType, 'p2wsh'> = 'p2pkh',
     network?: Network,
 ) {
     const calc = new TxWeightCalculator();
